Add class tests to ensure that memory manager & VM works correctly

+++ src/main/java/dev/xdark/ssvm/value/InstanceValue.java
@@ -77,8 +77,8 @@ public class InstanceValue extends ObjectValue {
 	 *
 	 * @return float value.
 	 */
-	public int getFloat(String field) {
-		return getMemoryManager().readInt(this, getFieldOffset(field, "F"));
+	public float getFloat(String field) {
+		return getMemoryManager().readFloat(this, getFieldOffset(field, "F"));
 	}
 
 	/**
